+++ apps/client/src/components/Search/Search.js
@@ -1080,6 +1080,7 @@ class Search extends React.PureComponent {
           searchModel={this.searchModel}
           searchOptions={searchOptions}
           updateSearchOptions={this.updateSearchOptions}
+          enabledSearchOptions={this.props.options.enabledSearchOptions}
           setSearchSources={this.setSearchSources}
           loading={loading}
           searchSources={searchSources}